Add expected results for stress test to compare against
Currently manually added the overall ingest/stress test speed to the
top of the file, would like to think of a more eloquent way of doing this within
the `bench.clj` functions down the line. Also, the sail version was
added manually and crux version info stripped from the JSON output.