Remove crux-bench-watdiv folder
Most of the content here has been ported over to `crux-bench` in some
way (certainly the contents of the bench and the watdiv-test), asides
from the scripts, dockerfiles, kubernetes folders and datomic properties folders.
For the most part, these have been replaced in some way, and I am
happy to cleanup what is here.