Create 'console-demo' project.
Starts up a read-only node which we can point people at so they can
play with the new crux console.

Using the james bond sample data (for now), submitting a set of
documents under the :bar entity as a showcase of entity history.

Contains all the necessary cloudformation files/build scripts to set
up the console on `console.crux.cloud`.