Can store + retrieve incomparable colls, #1001
Since #362 we sort the collections before we freeze them, to get a deterministic hash
Unfortunately this doesn't work if the collections aren't `Comparable`
So, if we can't sort the collections, we sort by the hashes of the elements, which is deterministic.
Much as we'd like to apply this across the board, we also take care not to change the hash of any data structure that was correctly hashed previously.