Improvements to behaviour of forms - remove vt/tt if not visible.
We were previously doing this in the form submit events based off of the value in
the db of show-vt/show-tt - this had issues, however, when these were
nil. In the subscription, we defer to whether or not they are present
in the URL, but the events did not have this behavior - rather than
adding these extra checks to the events, added a function which strips
non visible vt/tts before submitting values to the form events.