Ensuring that evict clears the bitemp range. Bump index version (#1358)
Fix mutability issue calling `all-keys-in-prefix` without eagerly evaluating the seq.
`all-keys-in-prefix` was doing a `kv/seek` on the provided iterator _before_ calling `lazy-seq`, which meant that multiple calls caused the shared iterator to be mutated concurrently.

This unfortunately requires an index version bump.
Users that re-use entity ids (i.e. put-evict-put) _are_ required to update
Otherwise, the impact of _not_ bumping are that evicted entities may still show in `entity-history` and `entity-tx`, but `entity` and `q` are unaffected.
We were correctly removing all personally identifiable entity data, but the above metadata remained.