replacing lein-sub with bash script - use `./lein-sub ...` instead of `lein sub ...`
fed up of working around it not setting CWD properly, and lein behaving differently if being run through `lein sub` or in the module directory itself.

too much in lein assumes it's being run from the root of the project - what happens is that lein's run from the directory above the module itself, so relative paths in config are one dir off.