fix a slightly more subtle JDBC deadlock caused by duplicate docs in the tx_events table:
it's possible that there are multiple docs with the same doc_id in the tx_events table - both through a put, evict and subsequent re-put, but also through a race condition between the `doc-exists?` check and the `insert-event!` call. when these were being updated, Postgres didn't always take the locks out for the multiple documents in the same order, causing the deadlock even when only one document was being inserted.

fix is to make the `doc-exists?` check explicitly take out a lock on the affected rows, in the same order between transactions. this does require splitting up the function by db-type, because MSSQL does `FOR UPDATE` differently, and SQLite doesn't have it at all