fix predicate variable resolution issue, resolves #1731 by splitting out variable existence checks from pred cycle checks:
this one used to work in 1.20.0 by virtue of the ordering of the checks happening to be the other way around. we don't need to check for cycles at that stage, though, as that's caught by the join-order calculation - it's enough to know that something else _could_ bind that pred input var if required.