Add arrow Interval support
We can read from all arrow interval vectors into a (arrow) PeriodDuration object, that represents a pair of java.time.Period and Duration (nanoseconds).

We support PeriodDuration as our canonical interval representation in memory, written to a IntervalMonthDayNanoVector. This **almost** preserves the entirety of representable values in all arrow & java & sql. But not quite.

The difference between years and months is lost when writing arrow, avoiding this will require a custom arrow vector - avoiding this for now.