Fixed issue where rename-step could rename columns in a `:scan`
e.g in select4 the tables contain columns starting with `x` e.g x1, x2 and so on. This introduced collisions with the generated name.

Fix is a custom postwalk. We could choose to introduce a different naming convention but that will require a lot of test fixes so avoiding that for now.