fix bug in `test-cast-to-ts` test caused by TZs being weird. resolves xtdb/core2#371:
In the test fixture, we cheat at comparing times of different types by 'just' converting everything to Instant and comparing that (we do it 'properly' in the EE)
error was in `#time/date "2033-03-27"` to TS in Asia/Beirut - calling 'at start of day' returns 01:00, because this is the day the clocks go forward, and there is no midnight that day in Beirut.
Spec says that to convert from date to ts is just zeroing everything, so we add a special case for date -> ts in the test that doesn't go via `->inst`.

Check postgres:

```
SET TIME ZONE 'Asia/Beirut';
SELECT DATE '2033-03-27'::timestamp; -- 2033-03-27 00:00:00
SELECT DATE '2033-03-27'::timestamptz; -- 2033-03-27 01:00:00+03:00
SELECT DATE '2033-03-27'::timestamptz::timestamp; -- 2033-03-27 01:00:00
```

Fun times!