Under contention await resize on miss to avoid runaway memory usage
Await the resize semaphore to apply back pressure on threads creating cache entries. We optimistically allow some (1024 as a hint) extra puts while a resize is under way, once your cache is full the resize loops will be running most of the time anyway, but the buffer may avoid some latency if demands on the cache are small.