Apply any unused conditions to outmost join in mega-join
This is a bit of a hack to get around the fact mega-join doesn't
currently explore all the possible join orders, therefore for
conditions that require more than 2 tables to be in scope it may not
produce a plan where the condition is valid.

That being said, its always valid (assuming our logical plans are
correct) to use any unused conditions as join conditions for the outmost
join.