Correct handling of Files/list in download-checkpoint error path.
This ensures that when a checkpoint destination exists that the check
for whether the destination has files can proceed
correctly. java.nio.files.Files/list returns a java.util.Stream<Path>
which cannot be coerced into a seq. By converting to an array the call
to (empty? ...) will succeed. In the case where there are a lot of
files, this might realise a large array into memory, but this seems
unlikely and the system is already in the process of crashing at this
point anyway, so doesn't seem too worrisome.