Add current-row-id cache
This cache was introduced to alleviate the cost of temporal resolution for
as of now queries. We do this by maintaining a clojure set of currently valid
row ids where app/sys time is now, and during transaction processing add and
remove row ids for changes for that affect "now".

However due to being a bitemp system app (valid) time moves forward irrespective
of new transactions coming in. To alleviate this we query the temporal index for
changes since the last time the cache was updated (latest transaction) and now,
and use those changes to adjust the current-row-id cache. We do this both on
every transaction (so that previously future changes are folded into the shared cache)
and on every scan (so that the cache can be used as long as the current time of the scan
after the tx-time of the query basis).

We rely on the current-row-id cache being a persistent data structure as we need the ability
to give versions of it to a watermark, this is so that we maintain the correctness of any given
scan/query even if the cache has subsequently been updated before that scan has been run.

Currently the current-row-id cache is not persisted to disk and so on startup a query from
the beginning of app-time is performed on the kd-tree to find the set of currently-valid row-ids.

There is an option to also maintain a sorted list of future changes in app-time so that we don't
have to query the kd-tree at all at query time and instead apply a subset of said changes based
on the current-time.