Use Reader in LiveHashTrie and Fix Table Watermark
Commit replaces the arrow vector in LiveHashTrie with a vector reader.
It also fixed a concurrency bug in LiveTable whereby if you had a
watermark out and the table was closed (for example because finish chunk
had been called on the Live Index) the buffers backing the iid vector in
the LiveHashTrie may be cleared from under you. To fix this issue we now
create a new LiveHashTrie with reference to a new retained iid vector
reader at the point a watermark is opened.