Use field accessor for latest-completed-tx to get dynamic ref
Closing over the symbol for a volatile-mutable field captures the static
value of that mutable field at the point the closure constructed.
Instead use of the field accessor delays resolving the value until the
accessor is invoked.

This fixes a concurrency bug in which latest->completed-tx could be
updated while await-tx-async was running. However because it has static
reference to the old value, it would create a future but fail to see
that the awaited transaction had subsequently been completed. This would
result in .awaitTxAsync blocking indefinately for a future that never
completes.