writers by name and name + field-type
This brings some consistency to the writers by
introducing a pattern for dealing with creation of
children writers of composite types. Most composite type writers now
have a method for creating child writers where one simply specifies
a name which by default will create a union type for the given child.
The second version for creating child writers takes a name and a
field-type. The given field-type is honoured if the given name does
not yet exist. A type mismatch error is thrown otherwise.

Fixing xtdb#2483 and xtdb#2418 .

Co-authored-by: James Henderson <james@jarohen.dev>
Co-authored-by: Finn Völkel <finn.volkel@gmail.com>