Fix ORDER BY when used in conjunction with select * in SQL
Commit removes recursive call to projected-columns from
column-references (via order-by-index), this is currently
only an issue for select * as in the process of expanding
the asterisk it calls back into column-references,
however my belief is that this mutually recursive relationship
between the 2 is unhelpful and column-references should always bottom out.

Result of this is that resolved-in-sort-key as a column-type has been
removed, instead we detect if an unqualified reference matches
a derived column during the errs attribute, rather than eagerly doing so
while resolving column references.