Add handling of `!Env` tags in YAML Decoder, test.
Done outside of the usual kotlinx.serialization code since it's more
of a pre-processing step - we'd otherwise have to ensure all decoders
for all values handled `!Env` somehow. Looked into kaml's `YamlNode`
but seemed like a bit of a dead end in terms of handling this
specifically. Seemed simplest to do a regex and handle it in that
was (similar in a way to how we use an #env reader for EDN). Tested
using mockk to fake out setting the environment variable.