Calculate expanded asterisk projected columns independently
Prior to this change, expand-asterisk (calculating the projected columns
as a result of using SELECT *) relied on projected-columns of the
table_primary. This had the side effect that any columns referenced
elsewhere in the query (such as in a WHERE clause) or columns not
included in * (such as temporal columns) were implicitly included if they
were present in the query. However the projected-columns attr of the
table_primary still needs to be aware of the extra columns that *
projects as this attribute is used to decide which columns need
including in scan, it does this by being dependant on expand-asterisk.

Additionally this commit changes how * interacts with group by. The
analyzer now first expands asterisks out into their underlying columns,
before checking them against the group by clause to see if thay can be
validly referenced in the select clause (or instead need adding to the
group by clause), and raising an error for any columns for which this is
not the case.