Only scan for known columns of a table when necessary
Previous commit moved the knowledge of what columns must be scanned
out as a result of using SELECT * to expand-asterisk. This commit
makes it so that those columns are only included in the projected
columns of a table primary (and therefore a scan) if there is an
asterisk present in the select_list of that query.