Refactor node to expose query phases
Operations such as parsing, preparing, binding and executing a query
feature in many SQL protocols and message flows. This commit takes the
opportunity to expose these as first class operations on an XTDB node
and various query objects.

The first new operation is prepareQuery, which is called on an XTDB
node, with an unparsed query and a set of query opts and returns and
after analysing and planning the query returns a PreparedQuery object.
It is at this point the structure of the database is captured (base tables,
the columns on those tables, and their respective types where relevant).
The owner of a PreparedQuery is assumed to be the caller, with it being
their responsibility to hold reference to and reuse as they see fit.

This PreparedQuery object exposes bind as a method, which when provided
a further set of query options (including params/args) will return a
BoundQuery object. It is at this point that the input data for a
query is captured (values of arguments/params to the query, which rows
in the base tables can be seen). The eventual result of executing a
BoundQuery should be deterministic. To allow for greater flexibility, if
the caller provides params as a query option, it is assumed they are
taking responsibility for opening and lifecycle of the param relation.
If instead the args option is provided, the BoundQuery takes
responsibility for opening and the param relation on the callers behalf and
will close said relation when the BoundQuery itself is closed.

A BoundQuery object exposes 2 main methods. The former is an optional
columnFields method, which returns an ordered list of the columnFields
which will be returned by the query. This can be thought of as
describing the result set of a given BoundQuery.

The latter is openCursor, which as the name suggests opens a cursor for
that query, which when stepped through returns the results of executing
the BoundQuery.

As of this commit prepareQuery, and therefore access to the subsequent
operations, is only available on the IXtdbInternal interface. These are
used by, and in turn re-exposed by, various internal clients to an XTDB
node such as pgwire and flightsql. Currently the only public query
operation remains openQueryAsync, which returns a stream of results,
however this has been reimplemented in terms of these lower level
operations.

Commit also renames IRaQuerySource to IQuerySource to better represent
it's new functionality as well as removes the caching as part of
qs.prepareRaQuery because as mentioned above, reuse of PreparedQueries
is the responsibility of the caller. However we may want to reinstate a
cache on the users behalf as part of openQueryAsync if there is a
sufficient performance benefit. Instead we now opt to cache planning of
queries as part of qs.planQuery.