Add support for describing prepared statements
Commit adds support for describing statements by optionally moving the
emission of a plan expression up to the columnFields method of a
preparedQuery. To allow for this to be possible the node requires callers
to supply types for all params when preparing a query, if they wish
to describe that query. A similar early binding of default-tz and
default-all-valid-time? happens.

In all the above cases if the user supplies of different value at bind
time (in the case of params, this would be a parameter of a different
type) we internally within a preparedQuery re-emit a new expression,
valid for the new set of values.

Additionally we also capture the state of any relevant schema, defined
as the total set of fields and their types from any tables referenced in
the query, when preparing a query. If at execution time this has
subsequently changed, we return an error instructing the user to
re-prepare their query.

Commit also includes reflecting these changes in the pgwire server,
which involved moving the prepare step from cmd-bind to cmd-parse,
adding a describe-statement cmd as well as mapping unknown param
and null types.

With regards to pg unknown types and untyped nulls, unlike xtdb postgres
lacks a standalone null type, nulls are always a value of a column/param
of some other type, inferred from context/schema. The type of null is
considered unknown. In the case of something like SELECT null, the
assumed behaviour is to return a column of type text with a null value.
We have opted to mirror this behaviour with one exception, that instead
of inferring the type of a null or unknown param from context we instead
assume its of type null, because as previously mentioned xtdb has first
class support for a null type.