Honor declared param types in pgwire
Commit ensures that we correctly honor param types in the instance
that multiple postgres types map onto a single XT type, such as
VARCHAR and TEXT both mapping onto UTF8. We opt to pick a single
postgres return type for the underlying XT type, however from a
pgwire perspective we should be seen to accept whichever type the
user specified.