Add support for pg ::regclass type and cast
Commit uses the newly available schema value available on
watermark to enable casting to and from an approximation
of the postgres regclass type. This type represents the OID
of a pg class (which is currently synonymous with a table in XTDB),
and uniquely identifies that relation via a hash of its namespaced
table name (schema-name + table-name). This makes it possible to
refer to tables by their name within SQL queries.

Commit does this by exposing the schema as an implicitly available variable
to the expression engine.