Minor fix in the SQL Quickstart docs
Hey! Hope you don't mind the drive-by PR for a minor fix in the docs. I'm happy send this some other way if you'd prefer. I'm working through the SQL Quickstart and noticed that the psql version of the "When did you know it?" section is missing the `INSERT` command. PS I like the SQL interface in the v2 changes. I saw yall at HYTRADBOI a few years ago but didn't get very far into testing it out at the time. Starting with SQL is easier for me.

Signed-off-by: Jack Provance <49460795+njprov@users.noreply.github.com>