Move responsibility for producing xt temporal period columns to BaseTable
Commit moves the responbility for producing valid/system_time columns
to base tables, as they only require special handling in that context.
This means elsewhere they can be treated as regular column references,
avoiding bugs such as #4054.

However for performance reasons we wish to continue to write the
optimised versions of the period predicate expressions. This is because
they are either more efficient, or are able to be pushed into scan
predicates in most cases. To enable this we write the inital expression
using the upper/lower period accessors and use a series of rules to
rewrite these expressions so they are both more performant and hopefully
able to be pushed into scan.