only taking a single L0 up to L1 so that all the nodes agree on what files to write
it's a bit different now that these are going to fluctuate in size - when you have files potentially decreasing in size,
the jobs will differ depending on what state each node sees when they start to compact

it's different when you know the files will only ever increase in size, because every node can use the size heuristic as an overestimate.

the difference with L0 is that we don't have metadata in those files (for speed of indexing) so we want to move those to L1 asap (at a cost of a double write, essentially)
whereas in L1+ we can afford to wait for a few.

the main thing is that we can't do 'take as many as you can', because this is the non-deterministic bit
i.e. you can say 'take 4, or however many take you over the size limit', because they'll all agree on that
but 'take as many as you can' they won't necessarily agree on