Add Microsecond Precision Intervals
Adds MonthDayMicro interval on top off mdn and uses it as the default.
Also adds some, but not complete support for casting between the durations/intervals etc.
of various precisions as code already existed for these and wanted to avoid removing features.
That being said, I think 99.9% of people are using micro durations and intervals,
so a case could be made for simplifying the logic/API here.