[bench] Yakbench use bench.stats
There's a bug in the implementation of Poisson on kixi.stats which essentially
means it caps at about 700, when lambda is around 700 e(-700) underflows to 0.0
so the knuth loop just returns a very small value. Implementations should switch
to using a normal variable for lambda > about 50.

Also in this commit is the addition of cycle for subs and user items, ensuring
our counts better match the original dataset.